# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SRA Resource Management System API
  version: 1.0.0
  description: >
    Business-layer API for the SRA Resource Management System (SRA-RMS).
    Provides CRUD over Clients, Projects, Resources, and Allocations, plus
    dashboard and reporting endpoints. Authentication is delegated to
    Microsoft Active Directory (OAuth2 / OpenID Connect); authorization is
    role-based (Administrator, General, Report).
  contact:
    name: SRA
    email: umesh.kodippili@sra.com.au
  license:
    name: Proprietary - SRA
servers:
  - url: https://api.rms.sra.com.au/v1
    description: Production
  - url: https://api.staging.rms.sra.com.au/v1
    description: Staging
  - url: http://localhost:5000/v1
    description: Local development
security:
  - oauth2: []
...
paths:
...
  /clients:
    get:
      tags: [Clients]
      summary: List clients
      operationId: listClients
      description: Roles - Administrator, General.
      parameters:
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: A paged list of clients.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientPage'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Clients]
      summary: Create a client
      operationId: createClient
      description: Roles - Administrator.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientCreate' }
      responses:
        '201':
          description: Client created.
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Client' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /clients/{clientId}:
    parameters:
      - $ref: '#/components/parameters/ClientId'
    get:
      tags: [Clients]
      summary: Get a client
      operationId: getClient
      description: Roles - Administrator, General. Includes projects and aggregated team.
      responses:
        '200':
          description: The client.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClientDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Clients]
      summary: Update a client
      operationId: updateClient
      description: Roles - Administrator.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientUpdate' }
      responses:
        '200':
          description: Client updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Client' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Clients]
      summary: Delete a client
      operationId: deleteClient
      description: >
        Roles - Administrator. Fails with 409 if the client has projects,
        unless cascade=true is supplied.
      parameters:
        - $ref: '#/components/parameters/Cascade'
      responses:
        '204': { description: Client deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /clients/{clientId}/projects:
    parameters:
      - $ref: '#/components/parameters/ClientId'
    get:
      tags: [Clients]
      summary: List a client's projects
      operationId: listClientProjects
      description: Roles - Administrator, General.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: A paged list of the client's projects.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------- Projects ----------------
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      description: Roles - Administrator, General.
      parameters:
        - $ref: '#/components/parameters/Search'
        - name: clientId
          in: query
          schema: { type: string, format: uuid }
        - name: billable
          in: query
          schema: { type: boolean }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ProjectStatus' }
        - name: startsAfter
          in: query
          schema: { type: string, format: date }
        - name: endsBefore
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: A paged list of projects.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Projects]
      summary: Create a project
      operationId: createProject
      description: Roles - Administrator.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201':
          description: Project created.
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /allocations:
    get:
      tags: [Allocations]
      summary: List allocations
      operationId: listAllocations
      description: Roles - Administrator, General. Filter by project, resource, or date range.
      parameters:
        - name: projectId
          in: query
          schema: { type: string, format: uuid }
        - name: resourceId
          in: query
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: A paged list of allocations.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Allocations]
      summary: Create an allocation
      operationId: createAllocation
      description: >
        Roles - Administrator. projectId and resourceId are supplied in the body.
        Validates dates against the project window; over-allocation returns
        warnings on the created object.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AllocationCreateFull' }
      responses:
        '201':
          description: Allocation created.
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Allocation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /allocations/{allocationId}:
    parameters:
      - $ref: '#/components/parameters/AllocationId'
    get:
      tags: [Allocations]
      summary: Get an allocation
      operationId: getAllocation
      description: Roles - Administrator, General.
      responses:
        '200':
          description: The allocation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Allocation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Allocations]
      summary: Update an allocation
      operationId: updateAllocation
      description: Roles - Administrator.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AllocationUpdate' }
      responses:
        '200':
          description: Allocation updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Allocation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Allocations]
      summary: Delete an allocation
      operationId: deleteAllocation
      description: Roles - Administrator.
      responses:
        '204': { description: Allocation deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------- Dashboard ----------------
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: Microsoft Active Directory / Entra ID (OpenID Connect).
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/common/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/common/oauth2/v2.0/token
          scopes:
            api://sra-rms/access_as_user: Access the SRA-RMS API as the signed-in user.
  headers:
    Location:
      description: URI of the newly created resource.
      schema: { type: string, format: uri }
  parameters:
    Search:
      name: q
      in: query
      description: Free-text search term.
      schema: { type: string }
    Page:
      name: page
      in: query
      description: 1-based page number.
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      description: Items per page.
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
    Sort:
      name: sort
      in: query
      description: 'Sort expression, e.g. "name" or "-startDate" (prefix - for descending).'
      schema: { type: string }
    Cascade:
      name: cascade
      in: query
      description: If true, delete dependents instead of returning 409.
      schema: { type: boolean, default: false }
...
    ClientId:
      name: clientId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    AllocationId:
      name: allocationId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: The request was invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication is required or has failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: The authenticated user lacks the required role.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: The request conflicts with the current state (e.g. duplicate or dependents exist).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
    PageMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }

    # ---- Enums ----
    ProjectStatus:
      type: string
      enum: [planned, active, onHold, completed, cancelled]
...
    EffortUnit:
      type: string
      enum: [hoursPerWeek, percent]

    # ---- Client ----
    ClientBase:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
    ClientCreate:
      allOf:
        - $ref: '#/components/schemas/ClientBase'
        - type: object
          required: [name]
    ClientUpdate:
      allOf:
        - $ref: '#/components/schemas/ClientBase'
        - type: object
          required: [name]
    Client:
      allOf:
        - $ref: '#/components/schemas/ClientBase'
        - type: object
          required: [id, name]
          properties:
            id: { type: string, format: uuid }
            projectCount: { type: integer }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    ClientDetail:
      allOf:
        - $ref: '#/components/schemas/Client'
        - type: object
          properties:
            projects:
              type: array
              items: { $ref: '#/components/schemas/Project' }
            team:
              type: array
              description: Distinct resources allocated across the client's projects.
              items: { $ref: '#/components/schemas/ResourceSummary' }
    ClientPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Client' }
        meta: { $ref: '#/components/schemas/PageMeta' }

    # ---- Project ----
    ProjectBase:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        code: { type: string, minLength: 1, maxLength: 50 }
        clientId: { type: string, format: uuid }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        budget: { type: number, format: double, minimum: 0 }
        remaining: { type: number, format: double, minimum: 0 }
        billable: { type: boolean, default: true }
        status: { $ref: '#/components/schemas/ProjectStatus' }
    ProjectCreate:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'
        - type: object
          required: [name, code, clientId, startDate, endDate]
...
    Project:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'
        - type: object
          required: [id, name, code, clientId, startDate, endDate]
          properties:
            id: { type: string, format: uuid }
            clientName: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    ProjectPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Project' }
        meta: { $ref: '#/components/schemas/PageMeta' }

    # ---- Resource ----
...
    ResourceSummary:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        primaryJobTitle: { type: string }
        imageUrl: { type: string, format: uri }
...
    AllocationBase:
      type: object
      properties:
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        effort: { type: number, format: double, minimum: 0 }
        effortUnit: { $ref: '#/components/schemas/EffortUnit' }
        roleOnProject: { type: string }
        billable: { type: boolean }
...
    AllocationCreateFull:
      description: Used on the top-level /allocations endpoint.
      allOf:
        - $ref: '#/components/schemas/AllocationBase'
        - type: object
          required: [projectId, resourceId, startDate, endDate, effort, effortUnit]
          properties:
            projectId: { type: string, format: uuid }
            resourceId: { type: string, format: uuid }
    AllocationUpdate:
      allOf:
        - $ref: '#/components/schemas/AllocationBase'
        - type: object
          required: [startDate, endDate, effort, effortUnit]
    Allocation:
      allOf:
        - $ref: '#/components/schemas/AllocationBase'
        - type: object
          required: [id, projectId, resourceId, startDate, endDate, effort, effortUnit]
          properties:
            id: { type: string, format: uuid }
            projectId: { type: string, format: uuid }
            projectName: { type: string }
            resourceId: { type: string, format: uuid }
            resourceName: { type: string }
            warnings:
              type: array
              description: Non-blocking validation warnings, e.g. over-allocation.
              items: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    AllocationPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Allocation' }
        meta: { $ref: '#/components/schemas/PageMeta' }

    # ---- Dashboard ----
...